Move the reads of `j2kt_web_experiment_enabled` closer to the j2cl_library rule
We're currently reading the attr deep in the J2CL rules even though not all
entrypoints pass this flag. We've generally avoided issues so far as the only
reads of the flag occur if `kt_compile` is being requested, which wouldn't be
the case for the `j2cl_proto` entrypoint.

However, when wiring up the Kotlin DSL protobuf generator I hit this problem. I
would either need to define this attribute on that rule, or just move the
reading of the flag closer to the rule entrypoint, as I've done here.

PiperOrigin-RevId: 764866865