Add an API for determining if an integration test is running under J2KT web
I reworked things a bit to split out the "environment" we're running in from
whether or not it's J2KT. This slightly reduces the amount of things that need
to be in the super-sourced Platform files.

Unfortunately determining whether J2KT Web is in use is tricky... we don't have
any marker for this. My initial approach was going to use a `select()` to swap
out the srcs to `j2cl_library`, but this doesn't work as we constantly iterate
over the srcs in the rule implementation, which you cannot do if the attribute
is being configured.

My workaround for this was to use a select in a `js_library` which will expose
the correct value which we will then pickup via jsinterop.

PiperOrigin-RevId: 766449264