[J2KT] Introduce J2ObjC / J2KT interop tests for `WeakReference` and `@Weak` / `@WeakOuter` annotations.
Currently, J2KT does not support `@Weak` and `@WeakOuter` annotations, which will result in retain-cycles on the boundary between ObjC and Kotlin/Native if cycle is not explicitly broken. This behavior is different from J2ObjC, which means that memory regression may occur when transitioning from J2ObjC to J2KT.

The `BUILD` file is updated so only `j2kt-native` and `j2objc` targets are tested, as other targets are irrelevant, and not all of them support `WeakReference`.

PiperOrigin-RevId: 778076278