[J2KT] Improve nullability propagation in RAW types, which has no effect in JDT frontend, but helps in Javac frontend.
With this CL, there are no more `null!!` occurrences in `NotNullAssertionProblems.java` readable when transpiled with Javac frontend.

The difference comes from the fact, that with the following declaration:

```
class Foo<T extends @Nullable Object> {
  Foo(T value) { ... }
}
```

the method descriptor of the RAW constructor call: `new Foo(null)` is `<init>(@Nullable Object)` in JDT, but `<init>(@NonNull Object)` in Javac.

PiperOrigin-RevId: 780042255