[KT] Cleanup extra nested blocks added from function inlining
When functions are inlined you may end up with one or two nested blocks. If
there's no parameters to the inline function then there's only one extra block,
otherwise you'll have an outer block to assign the parameters and then an inner
block for the body of the function.

We don't actually need these extra blocks and they make the output code more
verbose and difficult to read.

PiperOrigin-RevId: 781029874