[J2kt] new objc-names: reenable @ObjCName annotation for fields
Generate `@ObjCName` for fields even when the `NEW_OBJC_NAMES` flag is enabled. The field ObjCNames are generated not for compatibility with J2ObjC mangling but because there may be collisions between the generated properties (specifically their getter methods) and Java methods that have the same name as the property.

PiperOrigin-RevId: 782799204