Filter proto-referencing methods from J2ObjCHeaders. Except those with `@ObjectiveCName`
This is largely reverting cl 739987236.

When switching off `@ObjCName` generation in J2kt and injecting `@ObjCName` externally, the annotations for methods with proto parameters may change because J2kt doesn't have information about protos with custom Objective-C names, so J2kt might get the name of the wrapped method wrong. Methods with `@ObjectiveCName` in the Java source are safe because their name is known.

PiperOrigin-RevId: 785718832