[KT] Rewrite our name sanitization logic to escape invalid character.
The sanitization logic now follows the conservative rules of JsInteropRestrictionChecker. If, after sanitization, there's a conflict with an existing identifier, we'll take no further action, transpilation will fail, and the user will need to adjust their code.

We will follow-up to apply the same sanitization to field and class names.

PiperOrigin-RevId: 794468565