[javac] Update `JavaEnvironment.getTypeSubstitution()` to infer also from bounds of type variables.
Javac does not provide a way to retrieve inferred parameterizations directly;  `JavaEnvironment.getTypeSubstitution()` performs "unification" between the parameterized signature and the declaration to retrieve the values for the type parameters. The original code for this method is from errorprone, but it is only a best effort and is not complete.

PiperOrigin-RevId: 797787609