[KT] Use `undefined` to model Kotlin default arguments
The existing implementation we use comes from Kotlin/JVM wherein methods with
defaults always get a bridge method that takes a bitset to denote which
arguments were actually being passed.

In J2CL we can replace the bitset with just `undefined` as a marker, similar to
how you would model this in JS. However, we do not use JS's default initializer
syntax as this would change semantics, notably since they execute before the
body of the method, we lose trust that the class initializer has executed.

Another problem is that J2CL generally treats `null` and `undefined` as the same
thing allowing it to skip initialization of fields to null. Since we now need to
distinguish between the two, this requires coercing `undefined` to `null` at
call sites. For now only Kotlin callers require this as they're the only ones
that can call the bridge methods.

The implementation in this CL does several things:

 1. Removes the bitset from the default bridge methods.
 2. When a parameter is being omitted call sites will now explicitly pass in
    `undefined`. We cast this to the expected parameter type to avoid any
    autoboxing.
 3. When resolving default parameters in the bridge function we will check if
    the parameter `=== undefined`.
 4. Call sites will be updated to coerce `undefined` to `null` for any
    non-primitive type, unless `undefined` is being passed in literally.

Since all default methods and all callers of default methods are exclusively in
Kotlin, all changes to the AST are done in the Kotlin frontend. The only bit of
information carried in the J2CL AST is whether the parameter is optional or not.

PiperOrigin-RevId: 798260918