[JS] Used unboxed types for java.lang.Long
This mimics the existing behavior of Double, with the only tweak being that
instead of being backed by a JS primitive, it's backed by `goog.math.Long`.

Largely this is straight-forward, but some of the more interesting bits are:

 - `goog.math.Long` now needs to be included in the type union that forms
   Number/Comparable
 - A bunch of low-level APIs need to add `instanceof goog.math.Long` checks as
   we cannot differentiate based on the `typeof` result.

A minimal amount of client code needs to change, in particular places where
`java.lang.Long` are referred to from TS/JS directly. Those simply need to be
replaced by `goog.math.Long`.

PiperOrigin-RevId: 800486286