Update JsInterop restriction tests for jl.Long usages now that it's unboxed
Some of the tests were using `java.lang.Long` to trigger a warning previously,
these have either been deleted if it was explicitly testing `Long`, or replaced
by `Integer` if it was really testing other behavior, but just opted to `Long`
as its test value.

I've also added a test case for `Long` being used on an API to ensure that it
doesn't trigger an unusable-by-js warning.

PiperOrigin-RevId: 802200837