[KT] Resolve default parameters in local functions
Currently local functions with defaults are completely broken as we try to
invoke a bridge function that doesn't exist. Normally Kotlin hoists local
functions into the nearest declaration scope, but Kotlin/Closure keeps them
local and emits them as arrow functions. Therefore, we're not processing them
like other functions with defaults.

Instead of using a bridge we can instead add default parameter resolution into
the body of the function. Since the function isn't callable externally, we don't
need to worry about any overrides or Java callers, thus the bridge isn't
actually needed.

The only "trick" in this change is that we now allow the noted default dispatch
function to be the function itself. This will signal that the function has
defaults but intentionally does not have a bridge function.

PiperOrigin-RevId: 802677491