Remove the pattern of a type variable that is not referenced in the method signature for tests.
In many tests we had a pattern of defining a type variable just to be able to refer to it in the type body. This was useful to make intersection typed constructs e.g.

```
<T extends A & B> void testIntersection() {

  T ab = ...
  ....
}

```

This in general presents no problem, however when translated to kotlin by J2KT, the call should explicitly specify the type arguments. T

here is no general way to infer a type that satisfies the bounds, since it might not even exist (consider in the example above `A` and `B` to be interfaces without any class implementing both). In our tests there was always such a type and JDT would generally infer the correct one. javac on the other hand make no such attempt.

PiperOrigin-RevId: 803665026