[KT] Avoid reassigning optional parameters during default resolution
The JSCompiler cannot fold away parameters that are always set/unset if we both
need to read the parameter to check if it's undefined and then reassign it.

Instead we'll define new variables in the body of the function to hold the
result of the parameter or its default value, this way it's never actually
reassigned.

PiperOrigin-RevId: 805133632