Fix the mapping of KSuspentFunction{N} to KSuspendFunction.
Note that this mapping is most likely the root cause of this bug. We don't see the issue in the generated JavaScript code because we do not emit the type of the variable in JS as it is directly assigned but after this mapping there is a type mismatch in our ast. We will follow-up to fix that.

PiperOrigin-RevId: 805948909