[KT] Only coerce arguments at call sites for optional params
The existing logic wasn't validating that the parameter was actually optional
before applying coercion. We can just return the argument directly if it's
required.

The most visible change is that the dispatch receiver is no longer wrapped. This
is a bit confusing, but it's a side effect of the dispatch receiver being
included in the arguments now. Previously it was split out from valueArguments.

PiperOrigin-RevId: 807908783