Move coerceToNull/isUndefined to the nativebootstrap `Util` class
We plan on adding a JSCompiler pass to statically evaluate these calls, when
possible, and thus would like a stable place to have the implementations live.

However, because the calls needs to be referenced from the Kotlin frontend, we
need to proxy the calls through the `JsUtils` class as kotlinc has no way to
resolve the `Util` type.

PiperOrigin-RevId: 808687653