Lower call to the intrinsinc `coroutineContext` property getter.
The `coroutineContext` getter is an inline intrinsic function.
In Kotlin/JVM, this is replaced at code generation time. In Kotlin/JS, calls to this inline function are mapped to another top-level inline suspend function `getCoroutineContext`, which is then inlined.

J2CL cannot easily adopt the Kotlin/JS approach. The top-level replacement function would need to be attached to a FileClass,  which would require copying logic from the ExternalPackageParentPatcherLowering pass.
We opt to skip inlining the getter and instead lower the call later.

PiperOrigin-RevId: 808698270