[KT] fix default bridges not being marked as static
We'll reuse the `StaticDefaultFunctionLowering` pass from Kotlin/JVM to do this.
Unfortunately that lowering doesn't do anything about captured type parameters.
Thankfully the helper they use to convert a function to static does already have
the capability to copy captured type variables and remap references, we just
need to add a single line to pass them into the call.

Even more unfortunately, said helper function also has a bug where they don't
remap the return type. So we also need to copy the helper (and in turn the
private helpers it calls), so that we can make a single line patch to that as
well.

PiperOrigin-RevId: 810032601