[KT] Resolve types for intermediate type parameters
When we perform inlining or devirtualization of functions we need to resolve all
the type parameters present on the receiver. We currently mostly do this, but
if there multiple layers in the type hierarchy with type parameters assigned to
other type parameters, we incorrectly resolve to the intermediate type variable,
rather than the final type.

To correct this I now "flatten" the type mapping by going through each type
parameter and checking if it's assigned to another type parameter also in the
mapping. If so, we update to that value and repeat until we get to the final
value.

PiperOrigin-RevId: 811374700