Migrate the codebase to the new Kotlin IR parameter API, replacing deprecated ways of accessing function parameters and call arguments.
For more information about the new parameters API, please refer to this guide: http://google3/experimental/kmp/kotlin/docs/backend/IR_parameter_api_migration.md

This is a pre-work required for this migration to Kotlin 2.2.20. The deprecated APIs will require opt-in.

We did not change code that have been copied from JetBrains. The code will be  synced and automatically migrated to the new API during the migration to Kotlin 2.2.20.

PiperOrigin-RevId: 812254491