[javac] Add repro for the failure of the workaround for b/448907161.
The general workaround for missing type annotations in the type model in javac consists of traversing raw type attributes where all the annotations for a declaration are stored as a list of `TypeCompound`. The workaround started failing after a patch to jdk in the Java 9 time frame that changed the index of the superclass from -1 to 65335.

PiperOrigin-RevId: 814388842