[javac] Update workaround for retrieving type annotations for superclasses due to (old) changes to javac.
J2CL worksaround b/448907161 by looking at the raw type data of declarations. The raw type data is a list of `Compound` that contains annotation and positional information in the declaration. The `type_index` for the super class position changed from -1 to 65335 quite a long ago (in the Java 9) time frame, but the workaround was not updated.

PiperOrigin-RevId: 814412278