Always emit `J2ktInfo` providers if it's a J2KT-enabled build
Currently we are only emitting these providers if both (a) J2KT is enabled for
the build and (b) the target is explicitly enabled for J2KT. If it's missing,
then we on the fly J2KT import the `JavaInfo` provider from the `J2clInfo`
instead. If this target had J2KT dependencies, the `JavaInfo` would include
the `JavaInfo` emitted from the Kotlin/JVM build of those dependencies rather
than the original Java compilation.

In cases where a J2KT target has both J2KT and non-J2KT paths to a J2KT target,
we'll actually end up having a `JavaInfo` provider for both the original
pre-J2KT Java build of that target **and** the Kotlin/JVM compilation of the
J2KT output.

An example build graph of this would be:

```

               ┌──────────────┐
               │          J2KT│
               │     foo      │
               └──────┬───────┘
                      │
                      │
        ┌─────────────┴────────────┐
        │                          │
        │                          │
┌───────▼──────┐           ┌───────▼──────┐
│          J2CL│           │          J2KT│
│     bar      ┼───────────►     buzz     │
└──────────────┘           └──────────────┘
```

In this setup the J2KT transpilation of `foo` will have two jars for `buzz`, one
pre-J2KT and one post-J2KT. Which one actually gets used will solely depend upon
their order in the classpath.

This can ultimately cause problems as the post-J2KT jar will be missing
information (ex. nullability annotations) that `foo` needs.

To resolve this we now (almost) always emit a `J2ktInfo` provider if it's
enabled for the build. If the target is not J2KT enabled then we create a new
`JavaInfo` provider that uses the transitive `J2ktInfo` providers as
dependencies. This ensure J2KT builds always get the pre-J2KT `JavaInfo`
providers.

The only case I didn't handle is `j2cl_import`. Those targets really shouldn't
have J2KT dependencies so it's not particularly a concern.

PiperOrigin-RevId: 814823444