[KT] Package a srcjar for Kotlin standard library compilation ourselves
For some reason the srcjar that we get from the Kotlin/JVM compilation is not
being tolerated by kotlinc when we attempt to compile the builtins from sources.

I spent quite a bit of time trying to figure out where the discrepancy by
forcing file order to match a known good order, but I couldn't get it to pass.
This ultimately led me to believe that something about the file names is load
bearing and thus we're better off just packaging ourselves.

PiperOrigin-RevId: 817683331