Introduce character buffer to maintain character content.
Originally we have been using StringBuilder as a buffer to manage output. However StringBuilder emulation in J2CL is based on String concat operation. This normally performs really well for common usage since V8 highly optimized the String concat operations.

However when the majority of the concats are chars then the V8 CONCAT string look almost like a linked list of characters which behaves very poorly performance wise.

Due to this, CL creates a specialized char buffer implementation for the JS version of the minifier.

This improves the minifer performance by 3x for large projects.

PiperOrigin-RevId: 821775654