[KT] Sanitize Kotlin class names
Kotlin allows for way more names than would normally be tolerated by Java. To
ensure that we have reasonable names in the J2CL AST we'll sanitize the names in
the Kotlin frontend.

When a class/enum is a JsType/JsEnum we will _not_ sanitize the name as it's
implied that the user is requesting the type be exposed to JS as-is, unless
overridden by the `name` attribute on the annotation.

Note: the implementation for class is a bit different than how we handle this
for members. For members we only use the sanitized name if it's not a JsMember.
Otherwise we use the unsanitized name and rely upon JsInteropRestrictionChecker
to enforce.

For classes it's a bit more complicated as the name also impacts things like the
filename which is not otherwise impacted by the JS name. Therefore I went for an
approach where we always sanitize the class name, but for JsTypes we set the JS
name to the unsanitized name if none was specified. JsInteropRestrictionChecker
will then check that name.

From the user perspective either approach is the same: if a name needs to be
sanitized then the user must explicitly write it in the JsInterop annotation.

PiperOrigin-RevId: 823280207