Make `JsInteropRestrictionsCheckerTest` able to work with different frontends.
For most part `JsInteropRestrictionsCheckerTest` asserts contents of the error messages emitted but the transpiler itself which are independent from the frontend, however there are instances in which the frontend emits the error message and those can be quite different. This cl adds the ability to match based on a regex.

PiperOrigin-RevId: 830541713