[KT] Ignore property overrides that are coming from Java
Kotlin IR represents fields defined in a Java type as a IrProperty with a
backing field. If there's a Java subtype then the subtype will override the
IrProperty with its own field. This can cause issues for JsInterop as we'll see
the field defined multiple times in the hierarchy and believe there are name
collisions.

To avoid this we can simply ignore the property overrides from Java types.

PiperOrigin-RevId: 833912544