Use an output group to thread the output JAR that has test adapter artifacts
When J2KT is enabled for a test library the APT that generates the test adapter
runs first on the Java tree, and this works successfully. However, when J2KT
then runs, it doesn't bring forward those extra artifacts into the output JAR.
When we then later try to extract the resources, we fail to find them.

To resolve this we'll now emit a special output group and only for
`j2cl_generate_jsunit_suite` tagged libraries that contains either the pre-J2KT
JAR when J2KT is enabled, or just the regular output JAR.

This output group is then used to grab the JAR and pass it forward to the
extraction genrule.

PiperOrigin-RevId: 840016956