[KT]  Ensure interface clinit is called in default parameter bridges.
When a Kotlin interface has functions with default parameters, the compiler generates static bridge methods to handle the default value initialization. These bridge methods were previously marked as private, which prevented J2CL from inserting the necessary call to the interface's class initializer.

It the enclosing class of the bridge method is an interface, the bridge methods now inherits the visibility fro the original method.

PiperOrigin-RevId: 840404917