[javac] Emit Kythe metadata with root relative paths rather than source paths
We're currently emitting the source file path in the Kythe metadata which is
generally the path within the extract source jar output, not the original file
path.

I've copied the pattern we're using the JDT frontend where we switch between the
source path and target path based on whether Kythe metadata is being requested.

PiperOrigin-RevId: 844838076